Fix JASSIST-220
Default methods in Java 8 seems to mean that the
INVOKEINTERFACE opcode can sometimes refer to
an item that is a MethodrefInfo (instead of
an InterfaceMethodrefInfo).

This is now handled by only using the
ConstPool.getMethodref...() methods and casting
to MemberrefInfo in them (since MemberrefInfo
contains all the needed information).

This in turn means that all the
ConstPool.getInterfaceMethodref...() methods
can be removed.